\name{baseFrequencies}
\alias{baseFrequencies}
\alias{BF}
\title{Base Frequencies}
\description{
  This benchmark computes the base frequencies from five sequences
  stored in a FASTA file each with ten bases. Each sequence is made
  uniquely of A, C, G, T, or N. These values are compared with the
  values returned by the function \code{\link[ape]{base.freq}}.
}
\usage{
BF()
}
\author{Emmanuel Paradis}
\keyword{utilities}
