\name{BTIMES}
\alias{BTIMES}
\title{Branching Times}
\description{
  This benchmark assesses the branching times calculated on an
  ultrametric tree with five tips and four branching times equal to
  four, three, two, and one unit of time (from the root to the most
  recent node). The initial tree is stored in a Newick file. These
  values are compared with the values returned by the function
  \code{\link{branching.times}}.
}
\usage{
BTIMES()
}
\author{Emmanuel Paradis}
\keyword{utilities}
