\name{NJ_SaitouNei}
\alias{NJ_SaitouNei}
\title{Neiborgh-Joining}
\description{
  This benchmark assesses the neiborgh--joining (NJ) method of Saitou
  and Nei (1987) using the small example provided in their paper. This
  simple case results in a NJ tree with all branch lengths
  integers. This tree is compared with the tree returned by the function
  \code{\link{nj}}.
}
\usage{
NJ_SaitouNei()
}
\author{Emmanuel Paradis}
\references{
  Saitou, N. and Nei, M. (1987) The neighbor-joining method: a new
  method for reconstructing phylogenetic trees. \emph{Molecular Biology
    and Evolution}, \bold{4}, 406--425.
}
\keyword{utilities}
