\name{phylobench-package}
\alias{phylobench-package}
\alias{phylobench}
\docType{package}
\title{
Phylogenetic Benchmarks
}
\description{
  \pkg{phylobench} provides functions for testing phylogenetic functions.

  More information on \pkg{phylobench} can be found at
  \url{https://github.com/emmanuelparadis/phylobench}.
}

\author{
  Emmanuel Paradis

  Maintainer: Emmanuel Paradis <Emmanuel.Paradis@ird.fr>
}
\keyword{package}
